% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preFlopComp}
\alias{preFlopComp}
\title{Calculate pre-flop probability estimates}
\usage{
preFlopComp(your_cards, num_sims, num_folders, num_oppo)
}
\arguments{
\item{your_cards}{your cards as a length-4 character vector.}

\item{num_sims}{number of simulations.}

\item{num_folders}{number of folders.}

\item{num_oppo}{number of opponents.}
}
\value{
a character vector
}
\description{
Calculate pre-flop probability estimates
}
