% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preTurnComp}
\alias{preTurnComp}
\title{Calculate pre-turn probability estimates}
\usage{
preTurnComp(your_cards, num_sims, num_folders, num_oppo, comm_cards)
}
\arguments{
\item{your_cards}{your cards as a length-4 character vector.}

\item{num_sims}{number of simulations.}

\item{num_folders}{number of folders.}

\item{num_oppo}{number of opponents.}

\item{comm_cards}{community cards.}
}
\description{
Calculate pre-turn probability estimates
}
